sftp/requirements.txt
google-cloud-bigquery
pysftp